# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/subscriptions/create-by-type": {
      "post": {
        "operationId": "SubscriptionsController_createSubscriptionByType_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionCreateByTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/{id}": {
      "get": {
        "operationId": "SubscriptionsController_allSubscriptionById_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionsByIdDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Subscriptions"
        ]
      },
      "delete": {
        "operationId": "SubscriptionsController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionsByIdDto"
                }
              }
            }
          }
        },
        "tags": [
          "Subscriptions"
        ]
      }
    }
  },
  "info": {
    "title": "Custom Admin Panel API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "SubscriptionCreateByTypeDto": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "news",
              "blog"
            ]
          },
          "instagram": {
            "type": "string"
          }
        },
        "required": [
          "first_name",
          "email",
          "type"
        ]
      },
      "SubscriptionsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "news",
              "blog"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "nullable": true
          }
        },
        "required": [
          "id",
          "user_id",
          "type",
          "created_at",
          "updated_at",
          "is_active"
        ]
      },
      "SubscriptionsByIdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "news",
              "blog"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "is_active": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "type",
          "created_at",
          "updated_at",
          "is_active"
        ]
      }
    }
  }
}